Make ClusterManager own controller lifecycle
