+++ lib/iris/src/iris/cli/cluster.py
@@ -30,10 +30,9 @@ from connectrpc.errors import ConnectError
 from iris.client import IrisClient
 from iris.cluster.types import Entrypoint, ResourceSpec
 from iris.cluster.platform.cluster_manager import ClusterManager
-from iris.cluster.platform.controller_vm import ControllerVm
 from iris.config import make_local_config
 from iris.cluster.controller.slice_lifecycle import compute_slice_state_counts, slice_all_ready, slice_any_failed
-from iris.rpc import cluster_connect, cluster_pb2, config_pb2, vm_pb2
+from iris.rpc import cluster_connect, cluster_pb2, vm_pb2
 from iris.rpc.proto_utils import format_accelerator_display, vm_state_name
 from iris.time_utils import Timestamp
 
@@ -66,27 +65,6 @@ def _format_status_table(status: vm_pb2.AutoscalerStatus) -> str:
         )
         lines.append(line)
     return "\n".join(lines)
-
-
-def _resolve_group_zones(
-    group_config: config_pb2.ScaleGroupConfig,
-    platform_config: config_pb2.PlatformConfig,
-    zone_override: str | None,
-) -> list[str | None]:
-    if zone_override:
-        return [zone_override]
-
-    if platform_config.WhichOneof("platform") != "gcp":
-        return [None]
-
-    gcp = platform_config.gcp
-    if group_config.zones:
-        return list(group_config.zones)
-    if gcp.default_zones:
-        return list(gcp.default_zones)
-    if gcp.zone:
-        return [gcp.zone]
-    return [None]
 
 
 def _get_autoscaler_status(controller_url: str) -> vm_pb2.AutoscalerStatus:
@@ -235,48 +213,29 @@ def cluster_stop(ctx, zone: str | None):
     config = ctx.obj.get("config")
     if not config:
         raise click.ClickException("--config is required for cluster stop")
-    from iris.config import IrisConfig
+    manager = ClusterManager(config)
+    click.echo("Stopping controller and terminating slices...")
+    result = manager.stop_cluster(zone=zone)
 
-    iris_config = IrisConfig(config)
-    config = iris_config.proto
-    platform = iris_config.platform()
-    ops = platform
-    slice_ids_by_group: dict[str, dict[str | None, list[str]]] = {}
-
-    # Discover slices via platform ops (more reliable than RPC since controller may be down)
-    click.echo("Discovering existing slices via platform ops...")
-    for name, group_config in config.scale_groups.items():
-        for zone_name in _resolve_group_zones(group_config, config.platform, zone):
-            slices = ops.list_slices(group_config, zone=zone_name)
-            slice_ids = [s.slice_id for s in slices]
-            if slice_ids:
-                slice_ids_by_group.setdefault(name, {}).setdefault(zone_name, []).extend(slice_ids)
-            zone_label = zone_name or "default"
-            click.echo(f"  Found {len(slice_ids)} slice(s) in group '{name}' (zone={zone_label})")
-
-    ctrl = ControllerVm(platform, config)
-    click.echo("Stopping controller...")
-    try:
-        ctrl.stop()
-        click.echo("Controller stopped")
-    except Exception as e:
-        click.echo(f"Warning: Failed to stop controller: {e}", err=True)
-
-    if not slice_ids_by_group:
+    if not result.discovered:
         click.echo("No slices to terminate")
+        click.echo("Cluster stopped")
         return
 
-    total = sum(len(ids) for per_zone in slice_ids_by_group.values() for ids in per_zone.values())
-    click.echo(f"Terminating {total} slice(s)...")
-    for name, per_zone in slice_ids_by_group.items():
-        group_config = config.scale_groups[name]
+    total = sum(len(ids) for per_zone in result.discovered.values() for ids in per_zone.values())
+    click.echo(f"Discovered {total} slice(s)")
+
+    for name, per_zone in result.discovered.items():
         for zone_name, slice_ids in per_zone.items():
+            zone_label = zone_name or "default"
+            click.echo(f"  Group '{name}' (zone={zone_label}): {len(slice_ids)} slice(s)")
+
+            failures = result.failed.get(name, {}).get(zone_name, {})
             for slice_id in slice_ids:
-                try:
-                    ops.delete_slice(group_config, slice_id, zone=zone_name)
+                if slice_id in failures:
+                    click.echo(f"Failed to terminate {slice_id}: {failures[slice_id]}", err=True)
+                else:
                     click.echo(f"Terminated: {slice_id}")
-                except Exception as e:
-                    click.echo(f"Failed to terminate {slice_id}: {e}", err=True)
     click.echo("Cluster stopped")
